# docs/swagger.yaml
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse:
    properties:
      code:
        $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode'
      error:
        type: string
    type: object
...
  github_com_kotolino_lawyer_internal_handlers_responses.ErrorCode:
    enum:
    - INVALID_CREDENTIALS
    - UNAUTHORIZED
    - FORBIDDEN
    - EMAIL_NOT_VERIFIED
    - RESOURCE_NOT_FOUND
    - RESOURCE_ALREADY_EXISTS
    - INVALID_REQUEST
    - VALIDATION_FAILED
    - MISSING_REQUIRED_FIELD
    - INTERNAL_SERVER_ERROR
    - DATABASE_ERROR
    - OPERATION_FAILED
    - EMAIL_IS_ALREADY_IN_USE
    - EMAIL_IS_ALREADY_VERIFIED
    type: string
    x-enum-varnames:
    - ErrCodeInvalidCredentials
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeEmailNotVerified
    - ErrCodeResourceNotFound
    - ErrCodeResourceAlreadyExists
    - ErrCodeInvalidRequest
    - ErrCodeValidationFailed
    - ErrCodeMissingRequiredField
    - ErrCodeInternalServer
    - ErrCodeDatabaseError
    - ErrCodeOperationFailed
    - ErrCodeEmailIsAlreadyInUse
    - ErrCodeEmailIsAlreadyVerified
...
  github_com_kotolino_lawyer_internal_models.Attachment:
    properties:
      attachmentable_id:
        type: integer
      attachmentable_type:
        type: string
      created_at:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      uploaded_by:
        type: integer
    type: object
...
  github_com_kotolino_lawyer_internal_models.ChatMessage:
    properties:
      appointment_id:
        type: integer
      attachment:
        allOf:
        - $ref: '#/definitions/github_com_kotolino_lawyer_internal_models.Attachment'
        description: polymorphic one-to-one attachment
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      read:
        type: boolean
      receiver_id:
        type: integer
      sender_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  internal_handlers.ChatMessageRequest:
    properties:
      content:
        type: string
      receiver_id:
        type: integer
    required:
    - content
    - receiver_id
    type: object
...
  internal_handlers.attachmentBrief:
    properties:
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      id:
        type: integer
    type: object
  internal_handlers.chatResp:
    properties:
      appointment_id:
        type: integer
      attachment:
        $ref: '#/definitions/internal_handlers.attachmentBrief'
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      read:
        type: boolean
      receiver_id:
        type: integer
      sender_id:
        type: integer
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /appointments/{id}/attachments:
    post:
      consumes:
      - multipart/form-data
      description: Uploads and sends an attachment in a chat message
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Receiver ID
        in: formData
        name: receiver_id
        required: true
        type: integer
      - description: Attachment file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created message with attachment
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send attachment
      tags:
      - chat
  /appointments/{id}/messages:
    get:
      description: Gets all chat messages for a specific appointment
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.chatResp'
            type: array
        "400":
          description: Invalid appointment ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get chat messages
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Creates a new chat message for a specific appointment
      parameters:
      - description: Appointment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message details
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.ChatMessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_models.ChatMessage'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create chat message
      tags:
      - chat
...
  /messages/{id}:
    delete:
      description: Deletes a specific chat message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Message deleted
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid message ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete chat message
      tags:
      - chat
  /messages/{id}/read:
    post:
      description: Marks a specific chat message as read
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid message ID
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Mark message as read
      tags:
      - chat
  /messages/unread/count:
    get:
      description: Gets the count of unread messages for the current user
      produces:
      - application/json
      responses:
        "200":
          description: Count of unread messages
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_kotolino_lawyer_internal_handlers_responses.APIErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get unread message count
      tags:
      - chat
...
swagger: "2.0"